rename: `urlParsing` -> `parseUrl`
