+++ src/aics-image-viewer/shared/utils/parseUrl.ts
