+++ src/index.ts
@@ -1,6 +1,6 @@
 import ImageViewerApp from "./aics-image-viewer/components/App";
 
-export { addViewerParamsFromMessage, parseViewerUrlParams } from "./aics-image-viewer/shared/utils/urlParsing";
+export { addViewerParamsFromMessage, parseViewerUrlParams } from "./aics-image-viewer/shared/utils/parseUrl";
 export { writeMetadata, writeScenes } from "./aics-image-viewer/shared/utils/storage";
 
 export type {
